All activities except home screen functional
